+++ src/components/step-progress-bar/index.ts
@@ -1,7 +1,7 @@
 import Step from "./step";
 import StepProgressBar from "./step-progress-bar";
 
-export default {
+export {
     Step,
     StepProgressBar,
 };