+++ src/stories/select.stories.tsx
@@ -24,44 +24,68 @@ const items = [
 
 stories.add("Default", () => (
   <div>
-    <Select items={items} value={"id"} selected={"3"}/>
+    <Select items={items} valueKey={"id"} />
   </div>
 ));
 
-stories.add("Icon", () => (
-  <div>
-    <Select items={items} value={"id"} leftIcon={"close"} selected={"3"}/>
-    <br/>
-    <br/>
-    <Select items={items} value={"id"} rightIcon={"chevron-down"} selected={"3"}/>
-    <br/>
-    <br/>
-    {/* tslint:disable-next-line:max-line-length */}
-    <Select items={items} value={"id"} leftIcon={"filter"} leftIconColor={"primary"} rightIcon={"close"} selected={"3"}/>
-    <br/>
-    <br/>
+stories.add("General Usage", () => {
+  const [value, setValue] = React.useState("2");
+
+  const handleSelect = (e: any) => {
+    setValue(e.target.value);
+  };
+
+  const handleButton = () => {
+    setValue(value === "1" ? "2" : "1");
+  };
+
+  return (
     <div className={"flex"}>
-      {/* tslint:disable-next-line:max-line-length */}
-      <Select className={"select-w40"} items={items} value={"id"} leftIcon={"filter"} leftIconColor={"primary"} rightIcon={"close"} selected={"3"}/>
-      <Button className={"button-w60"} fluid>Hello World</Button>
+      <Select
+        className={"select-w40"}
+        items={items}
+        fluid
+        valueKey="id"
+        leftIcon="filter"
+        leftIconColor="primary"
+        rightIcon="close"
+        value={value}
+        onChange={handleSelect}
+      />
+      <Button className={"button-w60"} onClick={handleButton} fluid>Hello World</Button>
     </div>
-  </div>
-));
+  );
+});
+
+stories.add("Icon", () => {
+
+  return (
+    <div>
+      <Select items={items} valueKey="id" leftIcon={"close"} />
+      <br />
+      <br />
+      <Select items={items} valueKey="id" rightIcon={"chevron-down"} />
+      <br />
+      <br />
+      <Select items={items} valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" />
+    </div>
+  );
+});
 
 stories.add("Fluid", () => (
   <div>
-    <Select fluid items={items} rightIcon={"close"} selected={"3"}/>
+    <Select fluid items={items} rightIcon="close" />
   </div>
 ));
 
 stories.add("Disabled", () => (
   <div>
-    <Select disabled items={items} rightIcon={"close"} selected={"3"}/>
+    <Select disabled items={items} rightIcon="close" />
   </div>
 ));
 
 stories.add("Error", () => (
   <div>
-    <Select error items={items} rightIcon={"close"} selected={"3"}/>
+    <Select error items={items} rightIcon={"close"} />
   </div>
 ));